Fix DNS clean-up when TXT records are not created for all domains.
$toValidate is an array of domain indexes. The DNS clean-up code was
iterating over $toValidate, but using the loop variable to identify
domains instead of looking up the domain index in $toValidate.

If a certificate was requested for two domains, but the first was
already validated, a DNS record would only have been created for the
second domain. The clean-up code would then try and delete a
non-existent DNS record for the first domain instead of the record that
was created for the second domain.