Remove no longer needed try catches around calls to Get-Order. (#95)
Commit d505fae771b9eba0507154ccfdd2aa27ab7491ac changed Get-PAOrder so
that it returns $null instead of throwing an exception. The try catch
around the calls to Get-PAOrder in New-PAOrder and New-PACertificate can
therefore be removed.

This means that errors on refreshing the order (such as failures
connecting to the ACME server) will now be surfaced instead of being
silently caught and causing a new order or certificate to be created.