Fix broken DDNSNameserver param in RFC2136 plugin
The recent change to make the DDNSNameserver param optional in the RFC2136 plugin, actually disabled the ability to specify the DNS server manually, since the check to see if the param had been provided was using the wrong variable name. This forced nsupdate to always try to detect it via SOA, which could be wrong if people are using split DNS, etc.